Moved change into terminalInstance.ts
