+++ src/vs/workbench/parts/terminal/electron-browser/terminalActions.ts
@@ -655,40 +655,11 @@ export class RunActiveFileInTerminalAction extends Action {
 			this.notificationService.warn(nls.localize('workbench.action.terminal.runActiveFile.noFile', 'Only files on disk can be run in the terminal'));
 			return TPromise.as(void 0);
 		}
-		let uriPath: string = uri.fsPath;
-		const hasSpace = uriPath.indexOf(' ') !== -1;
-		if (isWindows) {
-			const exe = this.terminalService.getActiveInstance().shellLaunchConfig.executable;
-			const osVersion = (/(\d+)\.(\d+)\.(\d+)/g).exec(os.release());
-			let buildNumber: number = 0;
-			if (osVersion && osVersion.length === 4) {
-				buildNumber = parseInt(osVersion[3]);
-			}
-			// 17063 is the build number where wsl path was introduced.
-			// Update Windows uriPath to be executed in WSL.
-			if (((exe.indexOf('wsl') !== -1) || ((exe.indexOf('bash.exe') !== -1) && (exe.indexOf('git') === -1))) && (buildNumber >= 17063)) {
-				uriPath = 'runActive="$(wslpath ' + this._escapeNonWindowsPath(uriPath) + ')" && "${runActive}"';
-			} else if (hasSpace) {
-				uriPath = '"' + uriPath + '"';
-			}
-		} else if (!isWindows) {
-			uriPath = this._escapeNonWindowsPath(uriPath);
-		}
-		instance.sendText(uriPath, true);
-		return this.terminalService.showPanel();
-	}
 
-	private _escapeNonWindowsPath(path: string): string {
-		let newPath = path;
-		if (newPath.indexOf('\\') !== 0) {
-			newPath = newPath.replace(/\\/g, '\\\\');
-		}
-		if (!newPath && (newPath.indexOf('"') !== -1)) {
-			newPath = '\'' + newPath + '\'';
-		} else if (newPath.indexOf(' ') !== -1) {
-			newPath = newPath.replace(/ /g, '\\ ');
-		}
-		return newPath;
+		return instance.preparePathForTerminalAsync(uri.fsPath).then(path => {
+			instance.sendText(path, true);
+			return this.terminalService.showPanel();
+		});
 	}
 }
