+++ src/vs/workbench/parts/terminal/electron-browser/terminalInstance.ts
@@ -8,6 +8,7 @@ import * as nls from 'vs/nls';
 import * as platform from 'vs/base/common/platform';
 import * as dom from 'vs/base/browser/dom';
 import * as paths from 'vs/base/common/paths';
+import * as os from 'os';
 import { Event, Emitter } from 'vs/base/common/event';
 import { WindowsShellHelper } from 'vs/workbench/parts/terminal/node/windowsShellHelper';
 import { Terminal as XTermTerminal, ISearchOptions } from 'vscode-xterm';
@@ -651,6 +652,48 @@ export class TerminalInstance implements ITerminalInstance {
 				});
 			}
 		}
+	}
+
+	public preparePathForTerminalAsync(path: string): Promise<string> {
+		return new Promise<string>(c => {
+			let preparedPath = path;
+			const hasSpace = preparedPath.indexOf(' ') !== -1;
+			if (platform.isWindows) {
+				const exe = this.shellLaunchConfig.executable;
+				// 17063 is the build number where wsl path was introduced.
+				// Update Windows uriPath to be executed in WSL.
+				if (((exe.indexOf('wsl') !== -1) || ((exe.indexOf('bash.exe') !== -1) && (exe.indexOf('git') === -1))) && (TerminalInstance.getWindowsBuildNumber() >= 17063)) {
+					preparedPath = 'runActive="$(wslpath ' + this._escapeNonWindowsPath(preparedPath) + ')" && "${runActive}"';
+				} else if (hasSpace) {
+					preparedPath = '"' + preparedPath + '"';
+				}
+			} else if (!platform.isWindows) {
+				preparedPath = this._escapeNonWindowsPath(preparedPath);
+			}
+			c(preparedPath);
+		});
+	}
+
+	private _escapeNonWindowsPath(path: string): string {
+		let newPath = path;
+		if (newPath.indexOf('\\') !== 0) {
+			newPath = newPath.replace(/\\/g, '\\\\');
+		}
+		if (!newPath && (newPath.indexOf('"') !== -1)) {
+			newPath = '\'' + newPath + '\'';
+		} else if (newPath.indexOf(' ') !== -1) {
+			newPath = newPath.replace(/ /g, '\\ ');
+		}
+		return newPath;
+	}
+
+	public static getWindowsBuildNumber(): number {
+		const osVersion = (/(\d+)\.(\d+)\.(\d+)/g).exec(os.release());
+		let buildNumber: number = 0;
+		if (osVersion && osVersion.length === 4) {
+			buildNumber = parseInt(osVersion[3]);
+		}
+		return buildNumber;
 	}
 
 	public setVisible(visible: boolean): void {
