+++ src/vs/workbench/parts/terminal/electron-browser/terminalService.ts
@@ -221,14 +221,10 @@ export class TerminalService extends AbstractTerminalService implements ITermina
 		const is32ProcessOn64Windows = process.env.hasOwnProperty('PROCESSOR_ARCHITEW6432');
 		const system32Path = `${process.env['windir']}\\${is32ProcessOn64Windows ? 'Sysnative' : 'System32'}`;
 
-		const osVersion = (/(\d+)\.(\d+)\.(\d+)/g).exec(os.release());
 		let useWSLexe = false;
 
-		if (osVersion && osVersion.length === 4) {
-			const buildNumber = parseInt(osVersion[3]);
-			if (buildNumber >= 16299) {
-				useWSLexe = true;
-			}
+		if (TerminalInstance.getWindowsBuildNumber() >= 16299) {
+			useWSLexe = true;
 		}
 
 		const expectedLocations = {
